Move tl log forms
'group' attribute is now a hidden field since the name is displayed at the top
The selection of the attribute is confusing as the project was already chosen and a selection might lead to misclicking
